Extracts flow::executable to its own files
